Update CONTRIBUTING.md
Suggested an alternate way to download Homebrew, which becomes relevant in one of the possible ways to download an up-to-date version of Git on MacOS (under the "Clone the forked repository" section of the directions) because the current suggested way was gave a version of Homebrew that was to big and took up a lot of storage on Macs. Moved to my own fork. Fixed Link. (I am not sure if Bonnie wants this issue merged because she was questioning the necesity of the issue.)